# Remove FileStream long pinning, and simplify synchronization