@adamsitnik 2021-04-19
> If you could give them a go, that'd be helpful.

There are no regressions and the number of Gen 2 collections have dropped (as expected):

<details>

|                 Method |           Toolchain |  fileSize | userBufferSize |      options |          Mean | Ratio |   Gen 0 |  Gen 1 |  Gen 2 |   Allocated |
|----------------------- |-------------------- |---------- |--------------- |------------- |--------------:|------:|--------:|-------:|-------:|------------:|
|                   Read |  \after\corerun.exe |      1024 |           1024 |         None |      54.75 us |  1.00 |  0.4296 | 0.2148 |      - |     4,344 B |
|                   Read | \before\corerun.exe |      1024 |           1024 |         None |      54.52 us |  1.00 |  0.4496 | 0.4496 | 0.4496 |     4,346 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|                  Write |  \after\corerun.exe |      1024 |           1024 |         None |     434.29 us |  1.02 |       - |      - |      - |     4,344 B |
|                  Write | \before\corerun.exe |      1024 |           1024 |         None |     424.92 us |  1.00 |       - |      - |      - |     4,344 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|              ReadAsync |  \after\corerun.exe |      1024 |           1024 |         None |      69.71 us |  0.99 |  0.5708 | 0.2854 |      - |     5,088 B |
|              ReadAsync | \before\corerun.exe |      1024 |           1024 |         None |      70.71 us |  1.00 |  0.2880 | 0.2880 | 0.2880 |     5,088 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|             WriteAsync |  \after\corerun.exe |      1024 |           1024 |         None |     435.03 us |  1.01 |       - |      - |      - |     4,432 B |
|             WriteAsync | \before\corerun.exe |      1024 |           1024 |         None |     430.26 us |  1.00 |       - |      - |      - |     4,432 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|              ReadAsync |  \after\corerun.exe |      1024 |           1024 | Asynchronous |      82.80 us |  0.98 |  0.6614 | 0.3307 |      - |     5,256 B |
|              ReadAsync | \before\corerun.exe |      1024 |           1024 | Asynchronous |      84.13 us |  1.00 |  0.3324 | 0.3324 | 0.3324 |     5,240 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|             WriteAsync |  \after\corerun.exe |      1024 |           1024 | Asynchronous |     502.49 us |  1.00 |       - |      - |      - |     4,976 B |
|             WriteAsync | \before\corerun.exe |      1024 |           1024 | Asynchronous |     502.35 us |  1.00 |       - |      - |      - |     4,960 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|              OpenClose |  \after\corerun.exe |      1024 |              ? |         None |      43.76 us |  0.99 |       - |      - |      - |       224 B |
|              OpenClose | \before\corerun.exe |      1024 |              ? |         None |      44.36 us |  1.00 |       - |      - |      - |       224 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|             LockUnlock |  \after\corerun.exe |      1024 |              ? |         None |      84.72 us |  0.99 |       - |      - |      - |       224 B |
|             LockUnlock | \before\corerun.exe |      1024 |              ? |         None |      85.34 us |  1.00 |       - |      - |      - |       224 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|            SeekForward |  \after\corerun.exe |      1024 |              ? |         None |      55.00 us |  0.99 |       - |      - |      - |       224 B |
|            SeekForward | \before\corerun.exe |      1024 |              ? |         None |      55.59 us |  1.00 |       - |      - |      - |       224 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|           SeekBackward |  \after\corerun.exe |      1024 |              ? |         None |      58.61 us |  1.00 |       - |      - |      - |       224 B |
|           SeekBackward | \before\corerun.exe |      1024 |              ? |         None |      58.75 us |  1.00 |       - |      - |      - |       224 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|               ReadByte |  \after\corerun.exe |      1024 |              ? |         None |      55.96 us |  1.00 |  0.4464 | 0.2232 |      - |     4,344 B |
|               ReadByte | \before\corerun.exe |      1024 |              ? |         None |      56.03 us |  1.00 |  0.4433 | 0.4433 | 0.4433 |     4,346 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|              WriteByte |  \after\corerun.exe |      1024 |              ? |         None |     457.12 us |  1.00 |       - |      - |      - |     4,344 B |
|              WriteByte | \before\corerun.exe |      1024 |              ? |         None |     460.25 us |  1.00 |       - |      - |      - |     4,344 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|                  Flush |  \after\corerun.exe |      1024 |              ? |         None |   3,220.51 us |  0.97 |       - |      - |      - |     4,346 B |
|                  Flush | \before\corerun.exe |      1024 |              ? |         None |   3,314.11 us |  1.00 |       - |      - |      - |     4,346 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|             FlushAsync |  \after\corerun.exe |      1024 |              ? |         None |   6,217.04 us |  1.12 | 20.8333 |      - |      - |   275,067 B |
|             FlushAsync | \before\corerun.exe |      1024 |              ? |         None |   5,558.01 us |  1.00 | 20.8333 |      - |      - |   275,067 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|             CopyToFile |  \after\corerun.exe |      1024 |              ? |         None |     484.21 us |  1.00 |       - |      - |      - |     4,569 B |
|             CopyToFile | \before\corerun.exe |      1024 |              ? |         None |     486.41 us |  1.00 |       - |      - |      - |     4,569 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|        CopyToFileAsync |  \after\corerun.exe |      1024 |              ? |         None |     509.00 us |  0.98 |       - |      - |      - |     5,593 B |
|        CopyToFileAsync | \before\corerun.exe |      1024 |              ? |         None |     518.25 us |  1.00 |       - |      - |      - |     5,593 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|              OpenClose |  \after\corerun.exe |      1024 |              ? | Asynchronous |      45.66 us |  1.00 |       - |      - |      - |       264 B |
|              OpenClose | \before\corerun.exe |      1024 |              ? | Asynchronous |      45.71 us |  1.00 |       - |      - |      - |       264 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|             LockUnlock |  \after\corerun.exe |      1024 |              ? | Asynchronous |      89.37 us |  0.99 |       - |      - |      - |       264 B |
|             LockUnlock | \before\corerun.exe |      1024 |              ? | Asynchronous |      89.95 us |  1.00 |       - |      - |      - |       264 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|            SeekForward |  \after\corerun.exe |      1024 |              ? | Asynchronous |      57.77 us |  1.01 |       - |      - |      - |       264 B |
|            SeekForward | \before\corerun.exe |      1024 |              ? | Asynchronous |      57.21 us |  1.00 |       - |      - |      - |       264 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|           SeekBackward |  \after\corerun.exe |      1024 |              ? | Asynchronous |      62.40 us |  1.01 |       - |      - |      - |       264 B |
|           SeekBackward | \before\corerun.exe |      1024 |              ? | Asynchronous |      61.95 us |  1.00 |       - |      - |      - |       264 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|               ReadByte |  \after\corerun.exe |      1024 |              ? | Asynchronous |      74.92 us |  1.01 |  0.5952 | 0.2976 |      - |     4,826 B |
|               ReadByte | \before\corerun.exe |      1024 |              ? | Asynchronous |      74.32 us |  1.00 |  0.3005 | 0.3005 | 0.3005 |     4,809 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|              WriteByte |  \after\corerun.exe |      1024 |              ? | Asynchronous |     528.62 us |  1.01 |       - |      - |      - |     4,885 B |
|              WriteByte | \before\corerun.exe |      1024 |              ? | Asynchronous |     524.48 us |  1.00 |       - |      - |      - |     4,872 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|                  Flush |  \after\corerun.exe |      1024 |              ? | Asynchronous |  10,750.41 us |  1.00 |       - |      - |      - |   120,208 B |
|                  Flush | \before\corerun.exe |      1024 |              ? | Asynchronous |  10,795.36 us |  1.00 |       - |      - |      - |   119,588 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|             FlushAsync |  \after\corerun.exe |      1024 |              ? | Asynchronous |  10,757.45 us |  0.99 |       - |      - |      - |   160,661 B |
|             FlushAsync | \before\corerun.exe |      1024 |              ? | Asynchronous |  10,930.05 us |  1.00 |       - |      - |      - |   160,626 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|        CopyToFileAsync |  \after\corerun.exe |      1024 |              ? | Asynchronous |     539.89 us |  1.00 |       - |      - |      - |     6,352 B |
|        CopyToFileAsync | \before\corerun.exe |      1024 |              ? | Asynchronous |     541.66 us |  1.00 |       - |      - |      - |     6,336 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|                   Read |  \after\corerun.exe |   1048576 |            512 |         None |     731.20 us |  1.00 |       - |      - |      - |     4,344 B |
|                   Read | \before\corerun.exe |   1048576 |            512 |         None |     729.33 us |  1.00 |       - |      - |      - |     4,344 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|                  Write |  \after\corerun.exe |   1048576 |            512 |         None |   3,118.42 us |  1.01 |       - |      - |      - |     4,346 B |
|                  Write | \before\corerun.exe |   1048576 |            512 |         None |   3,185.80 us |  1.00 |       - |      - |      - |     4,346 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|              ReadAsync |  \after\corerun.exe |   1048576 |            512 |         None |   1,648.34 us |  1.00 |  6.9444 |      - |      - |    86,689 B |
|              ReadAsync | \before\corerun.exe |   1048576 |            512 |         None |   1,646.80 us |  1.00 |  6.9444 |      - |      - |    86,689 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|             WriteAsync |  \after\corerun.exe |   1048576 |            512 |         None |   3,828.83 us |  1.00 |       - |      - |      - |    78,208 B |
|             WriteAsync | \before\corerun.exe |   1048576 |            512 |         None |   3,915.84 us |  1.00 |       - |      - |      - |    78,211 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|              ReadAsync |  \after\corerun.exe |   1048576 |            512 | Asynchronous |   2,593.54 us |  1.11 |       - |      - |      - |    58,298 B |
|              ReadAsync | \before\corerun.exe |   1048576 |            512 | Asynchronous |   2,345.37 us |  1.00 |       - |      - |      - |    58,279 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|             WriteAsync |  \after\corerun.exe |   1048576 |            512 | Asynchronous |   5,342.20 us |  1.08 |       - |      - |      - |    50,052 B |
|             WriteAsync | \before\corerun.exe |   1048576 |            512 | Asynchronous |   4,983.02 us |  1.00 |       - |      - |      - |    50,020 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|                   Read |  \after\corerun.exe |   1048576 |           4096 |         None |     691.65 us |  1.00 |       - |      - |      - |       224 B |
|                   Read | \before\corerun.exe |   1048576 |           4096 |         None |     693.04 us |  1.00 |       - |      - |      - |       224 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|                  Write |  \after\corerun.exe |   1048576 |           4096 |         None |   3,597.98 us |  1.01 |       - |      - |      - |       226 B |
|                  Write | \before\corerun.exe |   1048576 |           4096 |         None |   3,596.56 us |  1.00 |       - |      - |      - |       226 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|              ReadAsync |  \after\corerun.exe |   1048576 |           4096 |         None |   1,085.40 us |  1.01 |       - |      - |      - |    29,321 B |
|              ReadAsync | \before\corerun.exe |   1048576 |           4096 |         None |   1,080.16 us |  1.00 |       - |      - |      - |    29,321 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|             WriteAsync |  \after\corerun.exe |   1048576 |           4096 |         None |   4,219.47 us |  1.03 |       - |      - |      - |    55,971 B |
|             WriteAsync | \before\corerun.exe |   1048576 |           4096 |         None |   4,105.06 us |  1.00 |       - |      - |      - |    55,971 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|              ReadAsync |  \after\corerun.exe |   1048576 |           4096 | Asynchronous |   2,125.01 us |  0.98 |       - |      - |      - |       929 B |
|              ReadAsync | \before\corerun.exe |   1048576 |           4096 | Asynchronous |   2,159.08 us |  1.00 |       - |      - |      - |       913 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|             WriteAsync |  \after\corerun.exe |   1048576 |           4096 | Asynchronous |   4,720.16 us |  1.01 |       - |      - |      - |    27,578 B |
|             WriteAsync | \before\corerun.exe |   1048576 |           4096 | Asynchronous |   4,714.79 us |  1.00 |       - |      - |      - |    27,562 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|       Read_NoBuffering |  \after\corerun.exe |   1048576 |          16384 |         None |     245.11 us |  0.99 |       - |      - |      - |       160 B |
|       Read_NoBuffering | \before\corerun.exe |   1048576 |          16384 |         None |     247.25 us |  1.00 |       - |      - |      - |       160 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|      Write_NoBuffering |  \after\corerun.exe |   1048576 |          16384 |         None |   2,889.89 us |  1.01 |       - |      - |      - |       162 B |
|      Write_NoBuffering | \before\corerun.exe |   1048576 |          16384 |         None |   2,897.49 us |  1.00 |       - |      - |      - |       162 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
|  ReadAsync_NoBuffering |  \after\corerun.exe |   1048576 |          16384 |         None |     393.51 us |  0.98 |       - |      - |      - |     7,664 B |
|  ReadAsync_NoBuffering | \before\corerun.exe |   1048576 |          16384 |         None |     399.94 us |  1.00 |       - |      - |      - |     7,664 B |
|                        |                     |           |                |              |               |       |         |        |        |             |
| WriteAsync_NoBuffering |  \after\corerun.exe |   1048576 |          16384 |         None |   3,104.27 us |  1.04 |       - |      - |      - |     7,666 B |
| WriteAsync_NoBuffering | \before\corerun.exe |   1048576 |          16384 |         None |   3,009.15 us |  1.00 |       - |      - |      - |     7,666 B |
…